// src/app/models/car.ts
export type Carro = {
    id: number
    vehicle: string
    volumetotal: number
    connected: number
    softwareUpdates: number
    vin: string
    img: string
}
export type VinInfos = {
    id: number
    odometro: number
    nivelCombustivel: number
    status: string
    lat: number
    long: number
}